Wrong prototype corrected.

+++ c/src/lib/libc/readdir.c
@@ -42,7 +42,7 @@ static char sccsid[] = "@(#)readdir.c	5.7 (Berkeley) 6/1/90";
 
 #include <dirent.h>
 
-long getdents(
+int getdents(
   int dd_fd,
   char *dd_buf,
   int dd_len
